Moved cardreader config from the interal package
