+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/payments/cardreader/CardReaderCountryConfigProvider.kt
@@ -1,6 +1,6 @@
 package com.woocommerce.android.ui.payments.cardreader
 
-import com.woocommerce.android.cardreader.internal.config.CardReaderConfigFactory
+import com.woocommerce.android.cardreader.config.CardReaderConfigFactory
 import javax.inject.Inject
 
 class CardReaderCountryConfigProvider @Inject constructor(
