+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/banner/BannerDisplayEligibilityCheckerTest.kt
@@ -9,9 +9,9 @@ import com.woocommerce.android.analytics.AnalyticsTracker.Companion.KEY_BANNER_P
 import com.woocommerce.android.analytics.AnalyticsTracker.Companion.KEY_BANNER_REMIND_LATER
 import com.woocommerce.android.analytics.AnalyticsTracker.Companion.KEY_BANNER_SETTINGS
 import com.woocommerce.android.analytics.AnalyticsTrackerWrapper
-import com.woocommerce.android.cardreader.internal.config.CardReaderConfigForCanada
-import com.woocommerce.android.cardreader.internal.config.CardReaderConfigForUSA
-import com.woocommerce.android.cardreader.internal.config.CardReaderConfigForUnsupportedCountry
+import com.woocommerce.android.cardreader.config.CardReaderConfigForCanada
+import com.woocommerce.android.cardreader.config.CardReaderConfigForUSA
+import com.woocommerce.android.cardreader.config.CardReaderConfigForUnsupportedCountry
 import com.woocommerce.android.tools.SelectedSite
 import com.woocommerce.android.ui.payments.banner.BannerDisplayEligibilityChecker
 import com.woocommerce.android.ui.payments.cardreader.CardReaderCountryConfigProvider
