+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/payments/cardreader/manuals/CardReaderManualsViewModelTest.kt
@@ -8,9 +8,9 @@ import com.woocommerce.android.cardreader.connection.SpecificReader
 import com.woocommerce.android.cardreader.connection.SpecificReader.Chipper2X
 import com.woocommerce.android.cardreader.connection.SpecificReader.StripeM2
 import com.woocommerce.android.cardreader.connection.SpecificReader.WisePade3
-import com.woocommerce.android.cardreader.internal.config.CardReaderConfigForCanada
-import com.woocommerce.android.cardreader.internal.config.CardReaderConfigForSupportedCountry
-import com.woocommerce.android.cardreader.internal.config.CardReaderConfigForUSA
+import com.woocommerce.android.cardreader.config.CardReaderConfigForCanada
+import com.woocommerce.android.cardreader.config.CardReaderConfigForSupportedCountry
+import com.woocommerce.android.cardreader.config.CardReaderConfigForUSA
 import com.woocommerce.android.initSavedStateHandle
 import com.woocommerce.android.ui.payments.cardreader.manuals.CardReaderManualsViewModel.ManualItem
 import com.woocommerce.android.viewmodel.BaseUnitTest
